requirements.txt
alpaca-py
pandas
pandas-ta
schedule
